+++ lib/string.h
@@ -1,18 +1,19 @@
+#include "stdint.h"
 void*
-memset(void *dst, int c, uint n);
+memset(void *dst, int c, uint8_t n);
 
 int
-memcmp(const void *v1, const void *v2, uint n);
+memcmp(const void *v1, const void *v2, uint8_t n);
 
 void*
-memmove(void *dst, const void *src, uint n);
+memmove(void *dst, const void *src, uint8_t n);
 
 void*
-memcpy(void *dst, const void *src, uint n);
+memcpy(void *dst, const void *src, uint8_t n);
 
 
 int
-strncmp(const char *p, const char *q, uint n);
+strncmp(const char *p, const char *q, uint8_t n);
 
 char*
 strncpy(char *s, const char *t, int n);
